Migrated IssueToken function to the flow package as this is really flow specific logic
